+++ CHANGES.md
@@ -1,1 +1,6 @@
 # Change Logs
+
+## v0.2.0
+
+- Allowed the synchronous callback for filter and messenger.
+- Added asynchronous client definition for messenger.